Refactored components to use global identity
