+++ src/components/IdentityForm.tsx
@@ -23,6 +23,22 @@ const IdentityForm: React.FC<Props> = ({ onSubmit, error }) => {
   const handleChange = (key: string, value: string) => {
     _setIdentity(prev => ({ ...prev, [key]: value }));
   };
+  const handleSubmitIdentity = () => {
+    if (_identity.displayName.trim() === '') {
+      setError('Display Name is required');
+      return;
+    }
+    setStage(1);
+    setChallenges(prev => ({
+      displayName: true,
+      ...Object.entries(identity).reduce((acc, [key, value]) => {
+        if (key !== 'displayName' && value.trim() !== '') {
+          acc[key] = { value: Math.random().toString(36).substring(2, 10), verified: false };
+        }
+        return acc;
+      }, {} as typeof prev)
+    }));
+  };
 
   const fieldNames: { [key: string]: string } = {
     displayName: 'Display Name',
@@ -58,7 +74,7 @@ const IdentityForm: React.FC<Props> = ({ onSubmit, error }) => {
       ))}
       {error && <p className="text-red-700 text-sm">{error}</p>}
       <button
-        onClick={onSubmit}
+        onClick={handleSubmitIdentity}
         className="mt-6 w-full bg-stone-700 hover:bg-stone-800 text-white py-2 text-sm font-semibold transition duration-300"
       >
         Submit
